made order of SNPs in output file constant based on the genomeID of the first snps file given. Stopped storing base before indel when that position is invariant.
Given the following alignment:

ref      T---AAA
seq1   TGGGAAA

Previously this was added to the dict used to store snp locations by adding the T from the reference sequence as well as the gaps. That way if the T was later found to be variant in another sequence it was already in the data structure.

This led to invariant positions in the final sequence. Now the T would only be added to the dict if it were later found to be variant in another sequence.

Also fixed an issue with alignments like the following being misrepresented by not correctly filling query sequence snp dicts with the variant sites found in other sequences.

ref     AA-GGG
seq1  A--GGG
seq2  AAAGGG

was misrepresented as

ref     AA-GGG
seq1  AA-GGG
seq2  AAAGGG

leading to incorrect sequences being output