changed snp counting for print statement
snp count was being done by counting the number of genome positions at the end of the script. If user uses -whole flag, this count is just the size of the reference genome + snps. Now it counts variant positions before filling in invariant sequence so the count is accurate either way.